requirements.txt
langchain
langchain_community
qdrant-client
tiktoken
langchain_groq
fastembed
python-dotenv